requirements.txt
python-dotenv==1.0.1
feedparser==6.0.11
nltk==3.8.1
pandas==2.2.3
pdfplumber==0.11.4
pymongo==4.9.1
rapidfuzz==3.9.6
Requests==2.32.3
